refactor(deepagents): move middleware merge helpers
